fix: move normalizeHastElement to server-only module
path.ts must be isomorphic (used by browser-side spa.inline.ts), but
normalizeHastElement imported formatting_improvement_html which pulls in
punctilio with Node.js built-ins. Moved to normalize-hast.ts so the
import chain stays server-side only.
